% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instabilityTest.R
\name{genWkSample}
\alias{genWkSample}
\title{genWkSample}
\usage{
genWkSample(fi, delta, sigma, kappa, a, b, k, N, sample = NULL)
}
\arguments{
\item{N}{\itemize{
\item размер выборки
}}

\item{sample}{\itemize{
\item выборка, ранее уже сгенерированная этой функцией.
}}
}
\description{
genWkSample
}
