% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instabilityTest.R
\name{getInstabilityQuantile}
\alias{getInstabilityQuantile}
\title{getInstabilityQuantile}
\usage{
getInstabilityQuantile(alpha, fi, delta, sigma, kappa, a, b, k)
}
\arguments{
\item{alpha}{\itemize{
\item уровень значимости
}}
}
\description{
getInstabilityQuantile
}
