% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{instabilityQuantile}
\alias{instabilityQuantile}
\title{instabilityQuantile}
\usage{
instabilityQuantile(alpha, w, fi, delta, sigma, kappa, a, b, k)
}
\arguments{
\item{alpha}{the significance level.}

\item{w}{initial state of Markov chain.}

\item{fi}{upper estimate for the absolute value of Markov chain increment in one step.}

\item{delta}{the downward drift that a process must exhibit in order to be stable.}

\item{sigma}{the lower estimate for the number of steps.}

\item{kappa}{the lower estimate for initial state.}

\item{a}{a parameter for time calculation. Time is calculated by a formula: a*(X_{i-1})+b.}

\item{b}{b parameter for time calculation. Time is calculated by a formula: a*(X_{i-1})+b.}

\item{k}{number of steps to select a new point.}
}
\value{
Quantile with given parameters
}
\description{
Calculation of a quantile with given parameters for a random variable stochastically dominating over a Markov chain.
}
