% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{instabilityQuantileTopEst}
\alias{instabilityQuantileTopEst}
\title{instabilityQuantileTopEst}
\usage{
instabilityQuantileTopEst(alpha, fi, delta, sigma, kappa, a, b, k)
}
\arguments{
\item{alpha}{\itemize{
\item уровень значимости
}}

\item{fi}{\itemize{
\item оценка сверху модуля приращения функции за 1 шаг.
}}

\item{delta}{\itemize{
\item отрицательный снос. В большинстве случаев можно задать равным 0.01 или 0.05
}}

\item{sigma}{\itemize{
\item оценка снизу числа шагов алгоритма поиска нестабильного лямбда (почти всегда можно брать 1)
}}

\item{kappa}{\itemize{
\item оценка снизу модуля начального состояния цепи
}}

\item{a}{\itemize{
\item множитель, определящий время следующей симуляции
}}

\item{b}{\itemize{
\item добавка, определяющая время следующей симуляции
}}

\item{k}{\itemize{
\item число шагов алгоритма.
}}
}
\description{
instabilityQuantileTopEst
}
