% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instabilityTest.R
\name{instabilityTest}
\alias{instabilityTest}
\title{instabilityTest}
\usage{
instabilityTest(w, fi, kappa, delta, tau)
}
\arguments{
\item{w}{increment of Markov chain in simulation time.}

\item{fi}{upper estimate for the absolute value of Markov chain increment in one step.}

\item{kappa}{lower estimate for the initial state of Markov chain.}

\item{delta}{the downward drift that a process must exhibit in order to be stable. Significance level analogue.}

\item{tau}{simulation time.}
}
\value{
TRUE, if  Markov chain is instable. FALSE doesn't mean that chain is stable.
}
\description{
Test simulation result on instability.
}
